Yet another checkstyle fix & relocate to the correct place
